@@ -14,7 +14,7 @@
         "drh/laravel-tanda": "^1.4",
         "fruitcake/laravel-cors": "^2.2",
         "guzzlehttp/guzzle": "^7.4",
-        "laravel/framework": "^9.28",
+        "laravel/framework": "^9.33",
         "laravel/sanctum": "^2.15",
         "laravel/tinker": "^2.7",
         "nabcellent/laravel-kyanda": "^2.0",
@@ -27,7 +27,7 @@
         "itsgoingd/clockwork": "^5.1",
         "laravel/pint": "^1.2",
         "mockery/mockery": "^1.5",
-        "nunomaduro/collision": "^6.2",
+        "nunomaduro/collision": "^6.3",
         "phpunit/phpunit": "^9.5.20",
         "spatie/laravel-ignition": "^1.0"
     },
