@@ -14,6 +14,6 @@ class FacadeServiceProvider extends ServiceProvider
      */
     public function register(): void
     {
-        $this->app->bind('localcarbon', fn () => new LocalCarbon);
+        $this->app->bind('localcarbon', fn() => new LocalCarbon);
     }
 }
