@@ -64,13 +64,14 @@ class TransactionRepository
             $debit_account = $data['method'] === PaymentMethod::MPESA ? $account['phone'] : $account['id'];
         }
 
-        $transactionsData = $transactions->map(fn ($t) => [
+        $transactionsData = $transactions->map(fn($t) => [
             'reference'   => $t->id,
             'product_id'  => $t->product_id,
             'amount'      => $t->amount,
             'destination' => $t->destination,
             'description' => $t->description,
         ]);
+
         $responseData = SidoohPayments::requestPayment($transactionsData, $data['method'], $debit_account);
 
         // TODO: Revert this to: if (!isset($response["data"]["payments"])) throw new Exception("Purchase Failed!");
@@ -82,7 +83,7 @@ class TransactionRepository
             throw new Exception('Purchase Failed!');
         }
 
-        $paymentData = array_map(function ($p) use ($responseData, $debit_account) {
+        $paymentData = array_map(function($p) use ($responseData, $debit_account) {
             return [
                 'transaction_id' => $p['reference'],
                 'payment_id'     => $p['id'],
@@ -151,7 +152,7 @@ class TransactionRepository
     {
         $responses = SidoohSavings::withdrawEarnings($transactions, $data['method']);
 
-        $transactions->each(function ($tx) use ($responses) {
+        $transactions->each(function($tx) use ($responses) {
             if (array_key_exists($tx->id, $responses['failed'])) {
                 $response = $responses['failed'][$tx->id];
 
@@ -193,7 +194,7 @@ class TransactionRepository
 
     public static function handleFailedPayments(Collection $transactions, Collection $failedPayments): void
     {
-        $transactions->each(function ($transaction) use ($failedPayments) {
+        $transactions->each(function($transaction) use ($failedPayments) {
             $transaction->payment->update(['status' => Status::FAILED]);
             $transaction->status = Status::FAILED;
             $transaction->save();
