@@ -18,7 +18,7 @@ class SidoohAccounts extends SidoohService
         $url = config('services.sidooh.services.accounts.url').'/accounts?with_user=true';
 
         //TODO: Process accounts to store with id for ease of fetching even one.
-        return Cache::remember('all_accounts', (60 * 60 * 24), fn () => parent::fetch($url));
+        return Cache::remember('all_accounts', (60 * 60 * 24), fn() => parent::fetch($url));
     }
 
     /**
@@ -30,7 +30,7 @@ class SidoohAccounts extends SidoohService
 
         $url = config('services.sidooh.services.accounts.url')."/accounts/$id?with_user=true";
 
-        $acc = Cache::remember($id, (60 * 60 * 24), fn () => parent::fetch($url));
+        $acc = Cache::remember($id, (60 * 60 * 24), fn() => parent::fetch($url));
 
         if (! $acc) {
             throw new Exception("Account doesn't exist!");
@@ -48,7 +48,7 @@ class SidoohAccounts extends SidoohService
 
         $url = config('services.sidooh.services.accounts.url')."/accounts/phone/$phone";
 
-        $acc = Cache::remember($phone, (60 * 60 * 24), fn () => parent::fetch($url));
+        $acc = Cache::remember($phone, (60 * 60 * 24), fn() => parent::fetch($url));
 
         if (! $acc) {
             throw new Exception("Account doesn't exist!");
@@ -74,7 +74,7 @@ class SidoohAccounts extends SidoohService
 
         $url = config('services.sidooh.services.accounts.url')."/accounts/$id/ancestors";
 
-        $ancestors = Cache::remember("${id}_ancestors", (60 * 60 * 24 * 28), fn () => parent::fetch($url));
+        $ancestors = Cache::remember("${id}_ancestors", (60 * 60 * 24 * 28), fn() => parent::fetch($url));
 
         if (! $ancestors) {
             throw new Exception("Account Ancestors don't exist!");
