@@ -13,7 +13,7 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('failed_jobs', function (Blueprint $table) {
+        Schema::create('failed_jobs', function(Blueprint $table) {
             $table->id();
             $table->string('uuid')->unique();
             $table->text('connection');
