@@ -13,7 +13,7 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('products', function (Blueprint $table) {
+        Schema::create('products', function(Blueprint $table) {
             $table->id();
 
             $table->string('name', 32)->unique();
