@@ -14,7 +14,7 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('transactions', function (Blueprint $table) {
+        Schema::create('transactions', function(Blueprint $table) {
             $table->id();
 
             $table->string('initiator', 20);
