@@ -13,7 +13,7 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('enterprises', function (Blueprint $table) {
+        Schema::create('enterprises', function(Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
             $table->json('settings');
