@@ -14,7 +14,7 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('payments', function (Blueprint $table) {
+        Schema::create('payments', function(Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('payment_id')->unique();
             $table->decimal('amount');
