@@ -28,8 +28,8 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function () {
-    Route::prefix('/products')->group(function () {
+Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function() {
+    Route::prefix('/products')->group(function() {
         Route::post('/airtime', AirtimeController::class);
         Route::post('/airtime/bulk', [AirtimeController::class, 'bulk']);
         Route::post('/utility', UtilityController::class);
@@ -40,11 +40,11 @@ Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function () {
 
 //        TODO: Should we have a similar endpoint for voucher purchase?
 //          Route::post('/voucher', WithdrawController::class);
-        Route::prefix('/vouchers')->group(function () {
+        Route::prefix('/vouchers')->group(function() {
             Route::post('/top-up', [VoucherController::class, 'topUp']);
         });
 
-        Route::prefix('/subscriptions')->group(function () {
+        Route::prefix('/subscriptions')->group(function() {
             Route::post('/', SubscriptionController::class);
         });
 
@@ -54,7 +54,7 @@ Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function () {
         Route::get('/subscription-types/default', SubscriptionTypeController::class);
     });
 
-    Route::prefix('/enterprises')->group(function () {
+    Route::prefix('/enterprises')->group(function() {
         Route::get('/', [EnterpriseController::class, 'index']);
         Route::post('/', [EnterpriseController::class, 'store']);
         Route::get('/{enterprise}', [EnterpriseController::class, 'show']);
@@ -63,20 +63,20 @@ Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function () {
         Route::post('/{enterprise}/accounts', [EnterpriseAccountController::class, 'store']);
     });
 
-    Route::prefix('/enterprise-accounts')->group(function () {
+    Route::prefix('/enterprise-accounts')->group(function() {
         Route::get('/{enterpriseAccount}', [EnterpriseAccountController::class, 'show']);
     });
 
-    Route::prefix('/payments')->group(function () {
+    Route::prefix('/payments')->group(function() {
         // Payments service callback
         Route::post('/callback', [PaymentsController::class, 'processCallback']);
     });
 
-    Route::prefix('/accounts')->group(function () {
+    Route::prefix('/accounts')->group(function() {
         Route::get('/airtime-accounts', [ProductController::class, 'getAllAirtimeAccounts']);
         Route::get('/utility-accounts', [ProductController::class, 'getAllUtilityAccounts']);
 
-        Route::prefix('/{accountId}')->group(function () {
+        Route::prefix('/{accountId}')->group(function() {
             Route::get('/details', [ProductController::class, 'getAccountDetails']);
 
             Route::get('/airtime-accounts', [ProductController::class, 'airtimeAccounts']);
@@ -88,7 +88,7 @@ Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function () {
         });
     });
 
-    Route::prefix('/savings')->group(function () {
+    Route::prefix('/savings')->group(function() {
         Route::post('/callback', [EarningController::class, 'processSavingsCallback']);
     });
 
@@ -120,13 +120,13 @@ Route::middleware('auth.jwt')->prefix('/v1')->name('api.')->group(function () {
 });
 
 // TODO: Research on how to secure or throttle unsecured callback endpoints
-Route::prefix('/v1')->name('api.')->group(function () {
-    Route::prefix('/products')->group(function () {
-        Route::prefix('/subscriptions')->group(function () {
+Route::prefix('/v1')->name('api.')->group(function() {
+    Route::prefix('/products')->group(function() {
+        Route::prefix('/subscriptions')->group(function() {
             Route::post('/check-expiry', [SubscriptionController::class, 'checkExpiry']);
         });
 
-        Route::prefix('/earnings')->group(function () {
+        Route::prefix('/earnings')->group(function() {
             Route::post('/save', [EarningController::class, 'saveEarnings']);
         });
     });
